refactor(mf): notepad - scss into css
