--- micro-frontends/src/mf-notepad/src/ui/modules/navigation/navigation.module.scss
@@ -1,11 +1,0 @@
-.navigation {
-  display: flex;
-  gap: 5px;
-  padding: 3px 2px;
-}
-
-.button {
-  @include font('medium', 11, 14);
-  color: var(--text-color);
-  cursor: pointer;
-}
